2014.11.20 Meeting - Version 220 Created
Version 220 (2.2.0) created in the process of troubleshooting the
Glassfish server error of "...220 doesn't have a public no args
constructor".  220 consists of a direct copy-and-paste of version 210
(in code text only, project was ground-up new), but has the same errors.
A default new project does not have these errors. The problem lies in
the code.