2015.01.29 Meeting Session and Home Session Cumulative Update
Fixed and added Mute/Unmute, Stop+Reset functionality to the [wavesurfer.js] music player after the first Play/Pause function.

For parts of the project reconsolidated, reordered code, deleted code that was no longer used and cluttering up classes.  To be continued further.  Notes need to be added.

And as always, minor improvements here and there.